feat: --precision arg now in the inference script

+++ hotshot_xl/pipelines/hotshot_xl_pipeline.py
@@ -962,7 +962,7 @@ class HotshotXLPipeline(DiffusionPipeline, FromSingleFileMixin, LoraLoaderMixin)
         video = []
         for frame_idx in tqdm(range(latents.shape[0])):
             video.append(self.vae.decode(
-                latents[frame_idx:frame_idx+1].to(dtype=self.vae.dtype)).sample)
+                latents[frame_idx:frame_idx+1]).sample)
         video = torch.cat(video)
         video = rearrange(video, "(b f) c h w -> b c f h w", f=video_length)
         video = (video / 2.0 + 0.5).clamp(0, 1)
